requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.47.0
django-bootstrap4==21.1
django-bootstrap5==21.2
django-ckeditor==6.2.0
django-clear-cache==0.3
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.9
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
